value_nをfloatからintに(floatの範囲を超えるため)

+++ source/engine/user-engine/mcts.h
@@ -10,7 +10,7 @@ class TreeConfig
 public:
 	float c_puct;
 	float play_temperature;
-	float virtual_loss;
+	int virtual_loss;
 	float value_scale;
 	bool clear_table_before_search;
 };
